Added FiniteAutomaton as state holder.
The FiniteAutomaton, as it is, simply contains a valid instance of a
finite automaton. It cannot perform any processing (can't check if it)
accepts a sequence.

Also added user uni-direction interaction (user can see the definition
of the finite automaton, but can't modify it - not at runtime. One can
modify it the automaton by re-writing the file specified in
Configuration.java)